pyproject.toml
[project]
name = "tupac"
version = "0.1.0"
description = "A GPT wrapper CLI with proper MCP tool & resource support"
readme = "README.md"
authors = [
    { name = "Tim Kellogg", email = "timothy.kellogg@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.7.0",
    "openai>=1.84.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "typer>=0.16.0",
]

[project.scripts]
tupac = "tupac.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.0.0",
]
